refactor: move cache to cached-exports & improve logs
